Fix dialog icon for low-level notification.
CFUserNotificationCreate is used to communicate with user during runAddKeyDialog. It requires an icon url, in our case we want to present the app icon. The app icon though is now managed using Asset Catalog that doesn’t provide urls to specific assets. That’s why there is now an additional resource called dialog_app_icon.png which is used for this notification.